Nexvigil_backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
motor>=3.3.2
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
uvloop>=0.19.0; sys_platform != 'win32'
httptools>=0.6.1; sys_platform != 'win32'
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0
slowapi>=0.1.9
aiofiles>=23.2.1
python-multipart>=0.0.9
opencv-python
ultralytics
psutil>=5.9.0

